// FinTracker.Client/src/types/models/ImportFileFormatViewModel.ts
type ImportFileFormatViewModel = {
    id?: number;
    importFileFormatName?: string;
    dateKey?: string;
    memoFormat?: string;
    amountKey?: string;
    invertAmounts?: boolean;
    headerLines?: number;
    delimiter?: string;
    image?: string;
};

export default ImportFileFormatViewModel;